refactor: onSuccess 바깥으로 빼기
