+++ src/pages/feed/edit/[id].tsx
@@ -1,10 +1,14 @@
-import { useMutation } from '@tanstack/react-query';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { useRouter } from 'next/router';
+import { playgroundLink } from 'playground-common/export';
 import { FC } from 'react';
 
 import { editFeed } from '@/api/endpoint/feed/editFeed';
-import { useGetPostQuery } from '@/api/endpoint/feed/getPost';
+import { getPost, useGetPostQuery } from '@/api/endpoint/feed/getPost';
+import { useGetPostsInfiniteQuery } from '@/api/endpoint/feed/getPosts';
 import AuthRequired from '@/components/auth/AuthRequired';
 import Loading from '@/components/common/Loading';
+import useEventLogger from '@/components/eventLogger/hooks/useEventLogger';
 import FeedUploadPage, { LoadingWrapper } from '@/components/feed/page/FeedUploadPage';
 import { FeedDataType } from '@/components/feed/upload/types';
 import useStringRouterQuery from '@/hooks/useStringRouterQuery';
@@ -13,11 +17,29 @@ import { setLayout } from '@/utils/layout';
 const FeedEdit: FC = () => {
   const { status, query } = useStringRouterQuery(['id'] as const);
   const { data } = useGetPostQuery(query?.id);
+  const editingId = data?.posts.id;
+  const router = useRouter();
+  const { logSubmitEvent } = useEventLogger();
+  const queryClient = useQueryClient();
 
   const { mutate, isPending } = useMutation({
     mutationFn: (requestBody: { data: FeedDataType; id: number | null }) =>
       editFeed.request({ postId: requestBody.id, ...requestBody.data }),
   });
+
+  const handleEditSubmit = ({ data, id }: { data: FeedDataType; id: number | null }) => {
+    mutate(
+      { data: data, id: id },
+      {
+        onSuccess: async () => {
+          logSubmitEvent('editCommunity');
+          queryClient.invalidateQueries({ queryKey: useGetPostsInfiniteQuery.getKey('') });
+          editingId && queryClient.invalidateQueries({ queryKey: getPost.cacheKey(`${editingId}`) });
+          await router.push(playgroundLink.feedList());
+        },
+      },
+    );
+  };
 
   if (isPending) {
     return (
@@ -50,7 +72,7 @@ const FeedEdit: FC = () => {
                 isBlindWriter: data.posts.isBlindWriter,
                 images: data.posts.images,
               }}
-              onSubmit={mutate}
+              onSubmit={handleEditSubmit}
               editingId={data.posts.id}
             />
           </AuthRequired>
