+++ src/pages/feed/upload.tsx
@@ -1,17 +1,38 @@
-import { useMutation } from '@tanstack/react-query';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { useRouter } from 'next/router';
+import { playgroundLink } from 'playground-common/export';
 import { FC } from 'react';
 
+import { useGetPostsInfiniteQuery } from '@/api/endpoint/feed/getPosts';
 import { uploadFeed } from '@/api/endpoint/feed/uploadFeed';
 import AuthRequired from '@/components/auth/AuthRequired';
 import Loading from '@/components/common/Loading';
+import useEventLogger from '@/components/eventLogger/hooks/useEventLogger';
 import FeedUploadPage, { LoadingWrapper } from '@/components/feed/page/FeedUploadPage';
 import { FeedDataType } from '@/components/feed/upload/types';
 import { setLayout } from '@/utils/layout';
 
 const FeedUpload: FC = () => {
+  const router = useRouter();
+  const { logSubmitEvent } = useEventLogger();
+  const queryClient = useQueryClient();
+
   const { mutate, isPending } = useMutation({
     mutationFn: (reqeustBody: { data: FeedDataType; id: number | null }) => uploadFeed.request({ ...reqeustBody.data }),
   });
+
+  const handlUploadSubmit = ({ data, id }: { data: FeedDataType; id: number | null }) => {
+    mutate(
+      { data: data, id: id },
+      {
+        onSuccess: async () => {
+          logSubmitEvent('submitCommunity');
+          queryClient.invalidateQueries({ queryKey: useGetPostsInfiniteQuery.getKey('') });
+          await router.push(playgroundLink.feedList());
+        },
+      },
+    );
+  };
 
   if (isPending) {
     return (
@@ -33,7 +54,7 @@ const FeedUpload: FC = () => {
           isBlindWriter: false,
           images: [],
         }}
-        onSubmit={mutate}
+        onSubmit={handlUploadSubmit}
       />
     </AuthRequired>
   );
